+++ packages/amplify-graphql-api-construct-tests/src/__tests__/additional-dependencies.ts
@@ -1,0 +1,1 @@
+export const authConstructDependency = '@aws-amplify/auth-construct@^1.0.0';