+++ packages/amplify-graphql-transformer/src/graphql-transformer.ts
@@ -33,6 +33,8 @@ import type {
 import { GraphQLTransform, ResolverConfig, UserDefinedSlot } from '@aws-amplify/graphql-transformer-core';
 import { Construct } from 'constructs';
 import { IFunction } from 'aws-cdk-lib/aws-lambda';
+import { GenerationTransformer } from '@aws-amplify/graphql-generation-transformer';
+import { ConversationTransformer } from '@aws-amplify/graphql-conversation-transformer';
 import { LogConfig } from 'aws-cdk-lib/aws-appsync';
 
 /**
@@ -62,6 +64,8 @@ export const constructTransformerChain = (options?: TransformerFactoryArgs): Tra
   const authTransformer = new AuthTransformer();
   const indexTransformer = new IndexTransformer();
   const hasOneTransformer = new HasOneTransformer();
+  const hasManyTransformer = new HasManyTransformer();
+  const belongsToTransformer = new BelongsToTransformer();
 
   // The default list of transformers should match DefaultDirectives in packages/amplify-graphql-directives/src/index.ts
   return [
@@ -71,10 +75,12 @@ export const constructTransformerChain = (options?: TransformerFactoryArgs): Tra
     new PredictionsTransformer(options?.storageConfig),
     new PrimaryKeyTransformer(),
     indexTransformer,
-    new HasManyTransformer(),
+    hasManyTransformer,
     hasOneTransformer,
     new ManyToManyTransformer(modelTransformer, indexTransformer, hasOneTransformer, authTransformer),
-    new BelongsToTransformer(),
+    belongsToTransformer,
+    new ConversationTransformer(modelTransformer, hasManyTransformer, belongsToTransformer, authTransformer),
+    new GenerationTransformer(),
     new DefaultValueTransformer(),
     authTransformer,
     new MapsToTransformer(),