+++ packages/amplify-e2e-tests/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.1](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-e2e-tests@3.22.11...amplify-category-api-e2e-tests@4.0.1) (2024-09-06)
+
+**Note:** Version bump only for package amplify-category-api-e2e-tests
+
+## [3.22.11](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-e2e-tests@3.22.10...amplify-category-api-e2e-tests@3.22.11) (2024-08-28)
+
+**Note:** Version bump only for package amplify-category-api-e2e-tests
+
 ## [3.22.10](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-e2e-tests@3.22.9...amplify-category-api-e2e-tests@3.22.10) (2024-08-20)
 
 ### Bug Fixes